Generic image size, fixed pixel input and edge write
Modified the program to take an image with size defined by variables
rather than hard coded.

Modified timing for pixel input so that the state machine will catch the
change on the next clock cycle, which we were missing before.

Modified the edge detected function, should detect when edge is high,
not just on o_valid.

Added some comments